// apps/frontend/app/api/v1/osograph/types/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
  [P in K]-?: NonNullable<T[P]>;
};
// ...
export type DataModel = {
  __typename?: "DataModel";
  createdAt: Scalars["DateTime"]["output"];
  dataset: Dataset;
  id: Scalars["ID"]["output"];
  isEnabled: Scalars["Boolean"]["output"];
  latestRelease?: Maybe<DataModelRelease>;
  latestRevision?: Maybe<DataModelRevision>;
  materializations: MaterializationConnection;
  modelContext?: Maybe<ModelContext>;
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  previewData: PreviewData;
  releases: DataModelReleaseConnection;
  revisions: DataModelRevisionConnection;
  runs: RunConnection;
  updatedAt: Scalars["DateTime"]["output"];
};

export type DataModelMaterializationsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DataModelReleasesArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DataModelRevisionsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DataModelRunsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
};
// ...
export type DataModelDependency = {
  __typename?: "DataModelDependency";
  alias?: Maybe<Scalars["String"]["output"]>;
  tableId: Scalars["ID"]["output"];
};

export type DataModelDependencyInput = {
  alias?: InputMaybe<Scalars["String"]["input"]>;
  dataModelId: Scalars["ID"]["input"];
};

export type DataModelEdge = {
  __typename?: "DataModelEdge";
  cursor: Scalars["String"]["output"];
  node: DataModel;
};

export const DataModelKind = {
  Full: "FULL",
  IncrementalByPartition: "INCREMENTAL_BY_PARTITION",
  IncrementalByTimeRange: "INCREMENTAL_BY_TIME_RANGE",
  IncrementalByUniqueKey: "INCREMENTAL_BY_UNIQUE_KEY",
  ScdType_2ByColumn: "SCD_TYPE_2_BY_COLUMN",
  ScdType_2ByTime: "SCD_TYPE_2_BY_TIME",
  View: "VIEW",
} as const;

export type DataModelKind = (typeof DataModelKind)[keyof typeof DataModelKind];
// ...
export type DataModelRelease = {
  __typename?: "DataModelRelease";
  createdAt: Scalars["DateTime"]["output"];
  dataModel: DataModel;
  dataModelId: Scalars["ID"]["output"];
  description?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  revision: DataModelRevision;
  revisionId: Scalars["ID"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};

export type DataModelReleaseConnection = {
  __typename?: "DataModelReleaseConnection";
  edges: Array<DataModelReleaseEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type DataModelReleaseEdge = {
  __typename?: "DataModelReleaseEdge";
  cursor: Scalars["String"]["output"];
  node: DataModelRelease;
};

export type DataModelRevision = {
  __typename?: "DataModelRevision";
  clusteredBy?: Maybe<Array<Scalars["String"]["output"]>>;
  code: Scalars["String"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  cron: Scalars["String"]["output"];
  dataModel: DataModel;
  dataModelId: Scalars["ID"]["output"];
  dependsOn?: Maybe<Array<DataModelDependency>>;
  description?: Maybe<Scalars["String"]["output"]>;
  end?: Maybe<Scalars["DateTime"]["output"]>;
  hash: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  kind: DataModelKind;
  kindOptions?: Maybe<DataModelKindOptions>;
  language: Scalars["String"]["output"];
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  partitionedBy?: Maybe<Array<Scalars["String"]["output"]>>;
  revisionNumber: Scalars["Int"]["output"];
  schema?: Maybe<Array<DataModelColumn>>;
  start?: Maybe<Scalars["DateTime"]["output"]>;
};

export type DataModelRevisionConnection = {
  __typename?: "DataModelRevisionConnection";
  edges: Array<DataModelRevisionEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type DataModelRevisionEdge = {
  __typename?: "DataModelRevisionEdge";
  cursor: Scalars["String"]["output"];
  node: DataModelRevision;
};

export type Dataset = {
  __typename?: "Dataset";
  createdAt: Scalars["DateTime"]["output"];
  creator: User;
  creatorId: Scalars["ID"]["output"];
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  /** Whether the current user's organization is subscribed to this dataset. */
  isSubscribed: Scalars["Boolean"]["output"];
  /**
   * The materializations for this dataset. Returns all materializations regardless of their source
   * (data model, static model, or data ingestion).
   */
  materializations: MaterializationConnection;
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  /**
   * The runs for this dataset. For USER_MODEL datasets, each run is related to individual data models.
   * FOR DATA_INGESTION datasets, each run is related to individual ingestion jobs.
   * For DATA_CONNECTION datasets, runs are not applicable and this field will be an empty list.
   */
  runs: RunConnection;
  tables: TableConnection;
  type: DatasetType;
  typeDefinition: DatasetTypeDefinition;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type MaterializationConnection = {
  __typename?: "MaterializationConnection";
  edges: Array<MaterializationEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type ModelContext = {
  __typename?: "ModelContext";
  columnContext?: Maybe<Array<ModelColumnContext>>;
  context?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  datasetId: Scalars["ID"]["output"];
  id: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
  tableId: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type Organization = {
  __typename?: "Organization";
  createdAt: Scalars["DateTime"]["output"];
  dataConnections: DataConnectionConnection;
  datasets: DatasetConnection;
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  members: UserConnection;
  name: Scalars["String"]["output"];
  notebooks: NotebookConnection;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type PreviewData = {
  __typename?: "PreviewData";
  isAvailable: Scalars["Boolean"]["output"];
  rows: Array<Scalars["JSON"]["output"]>;
};
// ...
export type RunConnection = {
  __typename?: "RunConnection";
  edges: Array<RunEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<
  TResult,
  TParent = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
  TArgs = Record<PropertyKey, never>,
> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<
  T = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
> = (
  obj: T,
  context: TContext,
  info: GraphQLResolveInfo,
) => boolean | Promise<boolean>;
// ...
export type DataModelResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataModel"] = ResolversParentTypes["DataModel"],
> = {
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  dataset?: Resolver<ResolversTypes["Dataset"], ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  isEnabled?: Resolver<ResolversTypes["Boolean"], ParentType, ContextType>;
  latestRelease?: Resolver<
    Maybe<ResolversTypes["DataModelRelease"]>,
    ParentType,
    ContextType
  >;
  latestRevision?: Resolver<
    Maybe<ResolversTypes["DataModelRevision"]>,
    ParentType,
    ContextType
  >;
  materializations?: Resolver<
    ResolversTypes["MaterializationConnection"],
    ParentType,
    ContextType,
    RequireFields<DataModelMaterializationsArgs, "first">
  >;
  modelContext?: Resolver<
    Maybe<ResolversTypes["ModelContext"]>,
    ParentType,
    ContextType
  >;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  previewData?: Resolver<
    ResolversTypes["PreviewData"],
    ParentType,
    ContextType
  >;
  releases?: Resolver<
    ResolversTypes["DataModelReleaseConnection"],
    ParentType,
    ContextType,
    RequireFields<DataModelReleasesArgs, "first">
  >;
  revisions?: Resolver<
    ResolversTypes["DataModelRevisionConnection"],
    ParentType,
    ContextType,
    RequireFields<DataModelRevisionsArgs, "first">
  >;
  runs?: Resolver<
    ResolversTypes["RunConnection"],
    ParentType,
    ContextType,
    RequireFields<DataModelRunsArgs, "first">
  >;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type DataModelReleaseResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataModelRelease"] = ResolversParentTypes["DataModelRelease"],
> = {
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  dataModel?: Resolver<ResolversTypes["DataModel"], ParentType, ContextType>;
  dataModelId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  description?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  revision?: Resolver<
    ResolversTypes["DataModelRevision"],
    ParentType,
    ContextType
  >;
  revisionId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
};
// ...
export type DataModelRevisionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataModelRevision"] = ResolversParentTypes["DataModelRevision"],
> = {
  clusteredBy?: Resolver<
    Maybe<Array<ResolversTypes["String"]>>,
    ParentType,
    ContextType
  >;
  code?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  cron?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  dataModel?: Resolver<ResolversTypes["DataModel"], ParentType, ContextType>;
  dataModelId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  dependsOn?: Resolver<
    Maybe<Array<ResolversTypes["DataModelDependency"]>>,
    ParentType,
    ContextType
  >;
  description?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  end?: Resolver<Maybe<ResolversTypes["DateTime"]>, ParentType, ContextType>;
  hash?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  kind?: Resolver<ResolversTypes["DataModelKind"], ParentType, ContextType>;
  kindOptions?: Resolver<
    Maybe<ResolversTypes["DataModelKindOptions"]>,
    ParentType,
    ContextType
  >;
  language?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  partitionedBy?: Resolver<
    Maybe<Array<ResolversTypes["String"]>>,
    ParentType,
    ContextType
  >;
  revisionNumber?: Resolver<ResolversTypes["Int"], ParentType, ContextType>;
  schema?: Resolver<
    Maybe<Array<ResolversTypes["DataModelColumn"]>>,
    ParentType,
    ContextType
  >;
  start?: Resolver<Maybe<ResolversTypes["DateTime"]>, ParentType, ContextType>;
};